Add BC6 support to nvtt lib and utils. - Use 3x3 eigensolver for initial fit in ZOH.  Slightly better perf and RMSE than power method. - Remove use of double precision in ZOH - speeds up by 12%. - Fixed RGBM encoding that was broken for HDR images. - Use gamma-2.0 space for RGBM for HDR images (improves precision in darks). - Use UNORM instead of TYPELESS formats when saving a DX10 .dds file.  The TYPELESS formats break most viewers. - Cleaned up warnings in ZOH code. - Command-line utils will warn if you give them an unrecognized parameter. - Added VS2010 profiling results.

+++ trunk/src/nvtt/bc6h/utils.h
@@ -47,8 +47,8 @@ public:
     static ::Format FORMAT;     // this is a global -- we're either handling unsigned or unsigned half values
 
     // error metrics
-    static double norm(const nv::Vector3 &a, const nv::Vector3 &b);
-    static double mpsnr_norm(const nv::Vector3 &a, int exposure, const nv::Vector3 &b);
+    static float norm(const nv::Vector3 &a, const nv::Vector3 &b);
+    static float mpsnr_norm(const nv::Vector3 &a, int exposure, const nv::Vector3 &b);
 
     // conversion & clamp
     static int ushort_to_format(unsigned short input);
